basic production done. More modification can be needed

+++ database/migrations/2019_03_20_195650_add_colmn_user_id_buyers_table.php
@@ -14,7 +14,7 @@ class AddColmnUserIdBuyersTable extends Migration
     public function up()
     {
         Schema::table('buyers', function (Blueprint $table) {
-            $table->integer('user_id')->default(0)->nullable()->index();
+            $table->bigInteger('user_id')->default(0)->nullable()->index();
         });
     }
